requirements.txt
astroid==2.5.3
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
docutils==0.17
html5lib==1.1
idna==2.10
isort==5.8.0
Kivy==2.0.0
Kivy-Garden==0.1.4
#kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
pafy==0.5.5
Pillow==8.2.0
pycodestyle==2.7.0
pygame==2.0.1
Pygments==2.8.1
pylint==2.7.4
python-vlc==3.0.12117
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.4
webencodings==0.5.1
wrapt==1.12.1
youtube-dl==2021.4.7
